backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
web3==6.15.0
requests==2.31.0
python-dotenv==1.0.1
pandas>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0